Move MultiByteBufferInputStream out of it's own package/directory.
